+++ src/main/java/callprotector/spring/domain/user/controller/TestController.java
@@ -1,4 +1,4 @@
-package callprotector.spring.web.controller;
+package callprotector.spring.domain.user.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
