+++ src/main/java/callprotector/spring/domain/user/entity/User.java
@@ -1,5 +1,6 @@
-package callprotector.spring.domain;
+package callprotector.spring.domain.user.entity;
 
+import callprotector.spring.domain.ChatSession;
 import callprotector.spring.global.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
