+++ src/main/java/callprotector/spring/domain/user/entity/VerificationToken.java
@@ -1,11 +1,9 @@
-package callprotector.spring.domain;
+package callprotector.spring.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDateTime;
-import java.util.Random;
-import java.util.UUID;
 
 @Entity
 @Getter
